rework unit tests
