+++ OLT.Extensions.Swagger.sln
@@ -3,15 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31919.166
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OLT.Extensions.SwaggerGen.Tests", "tests\OLT.Extensions.SwaggerGen.Tests\OLT.Extensions.SwaggerGen.Tests.csproj", "{45C43A8B-FEB2-4AA9-9AF0-AF8936BE7663}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{4457799F-3AF6-407B-B1DF-8CB6C1095AA9}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{E20F2D2C-A1E2-4655-BE71-9C858B2078AC}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OLT.Extensions.SwaggerGen.Versioning", "src\OLT.Extensions.SwaggerGen.Versioning\OLT.Extensions.SwaggerGen.Versioning.csproj", "{03BA13FE-4FDC-4E27-B64F-DCF1BE3A7FE2}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OLT.AspNetCore.Versioning", "tests\OLT.AspNetCore.Versioning\OLT.AspNetCore.Versioning.csproj", "{A50C8A16-9229-42EC-B4E2-2B3E044C4ADE}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".build", ".build", "{E7350367-380A-4D25-9BFE-DCA0DD03B027}"
 	ProjectSection(SolutionItems) = preProject
@@ -20,32 +16,29 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".build", ".build", "{E73503
 		.github\workflows\release.yml = .github\workflows\release.yml
 	EndProjectSection
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OLT.Extensions.SwaggerGen.Versioning.Tests", "tests\OLT.Extensions.SwaggerGen.Versioning.Tests\OLT.Extensions.SwaggerGen.Versioning.Tests.csproj", "{FBD3714F-CB20-4885-A894-1CA679E7A06B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{45C43A8B-FEB2-4AA9-9AF0-AF8936BE7663}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{45C43A8B-FEB2-4AA9-9AF0-AF8936BE7663}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{45C43A8B-FEB2-4AA9-9AF0-AF8936BE7663}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{45C43A8B-FEB2-4AA9-9AF0-AF8936BE7663}.Release|Any CPU.Build.0 = Release|Any CPU
 		{03BA13FE-4FDC-4E27-B64F-DCF1BE3A7FE2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{03BA13FE-4FDC-4E27-B64F-DCF1BE3A7FE2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{03BA13FE-4FDC-4E27-B64F-DCF1BE3A7FE2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{03BA13FE-4FDC-4E27-B64F-DCF1BE3A7FE2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A50C8A16-9229-42EC-B4E2-2B3E044C4ADE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A50C8A16-9229-42EC-B4E2-2B3E044C4ADE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A50C8A16-9229-42EC-B4E2-2B3E044C4ADE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A50C8A16-9229-42EC-B4E2-2B3E044C4ADE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FBD3714F-CB20-4885-A894-1CA679E7A06B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FBD3714F-CB20-4885-A894-1CA679E7A06B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FBD3714F-CB20-4885-A894-1CA679E7A06B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FBD3714F-CB20-4885-A894-1CA679E7A06B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{45C43A8B-FEB2-4AA9-9AF0-AF8936BE7663} = {E20F2D2C-A1E2-4655-BE71-9C858B2078AC}
 		{03BA13FE-4FDC-4E27-B64F-DCF1BE3A7FE2} = {4457799F-3AF6-407B-B1DF-8CB6C1095AA9}
-		{A50C8A16-9229-42EC-B4E2-2B3E044C4ADE} = {E20F2D2C-A1E2-4655-BE71-9C858B2078AC}
+		{FBD3714F-CB20-4885-A894-1CA679E7A06B} = {E20F2D2C-A1E2-4655-BE71-9C858B2078AC}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {26FC79E1-445C-4094-A33A-45AE144022ED}
