--- tests/OLT.AspNetCore.Versioning/GlobalUsings.cs
@@ -1,1 +1,0 @@
-global using Xunit;
