--- tests/OLT.Extensions.SwaggerGen.Tests/OLT.Extensions.SwaggerGen.Tests.csproj
@@ -1,38 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<TargetFramework>net9.0</TargetFramework>
-		<Nullable>enable</Nullable>
-		<IsPackable>false</IsPackable>
-		<SonarQubeExclude>true</SonarQubeExclude>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Faker.Net" Version="2.0.163" />
-		<PackageReference Include="FluentAssertions" Version="6.12.2" />
-		<PackageReference Include="Microsoft.AspNetCore.TestHost" Version="9.0.0" />
-		<PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="9.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Configuration" Version="9.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="9.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="9.0.0" />
-		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.11.1" />
-		<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-		<PackageReference Include="xunit" Version="2.9.2" />
-		<PackageReference Include="Xunit.DependencyInjection" Version="9.6.0" />
-		<PackageReference Include="xunit.runner.visualstudio" Version="2.8.2">
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-			<PrivateAssets>all</PrivateAssets>
-		</PackageReference>
-		<PackageReference Include="coverlet.collector" Version="6.0.2">
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-			<PrivateAssets>all</PrivateAssets>
-		</PackageReference>
-	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\..\src\OLT.Extensions.SwaggerGen.Versioning\OLT.Extensions.SwaggerGen.Versioning.csproj" />
-	</ItemGroup>
-
-
-
-</Project>
