--- tests/OLT.Extensions.SwaggerGen.Tests/Startup.cs
@@ -1,15 +1,0 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using OLT.Extensions.SwaggerGen.Versioning;
-
-namespace OLT.Extensions.SwaggerGen.Tests
-{
-
-    //This is needed for xUnit DI
-    public class Startup
-    {
-       
-    }
-}
