+++ tests/OLT.Extensions.SwaggerGen.Versioning.Tests/Controllers/Models/RouteIdParams.cs
@@ -1,4 +1,4 @@
-﻿namespace OLT.Extensions.SwaggerGen.Tests
+﻿namespace OLT.Extensions.SwaggerGen.Versioning.Tests.Controllers.Models
 {
     public class RouteIdParams
     {
