+++ tests/OLT.Extensions.SwaggerGen.Versioning.Tests/Startup.cs
@@ -1,0 +1,8 @@
+﻿namespace OLT.Extensions.SwaggerGen.Versioning.Tests
+{
+    //This is needed for xUnit DI
+    public class Startup
+    {       
+
+    }
+}
